+++ src/task/LongTimerTask.ts
@@ -2,7 +2,7 @@ import { DurableOrchestrationContext, TimerTask } from "durable-functions";
 import { WhenAllTask } from "./WhenAllTask";
 import moment = require("moment");
 import { CreateTimerAction } from "../classes";
-import { TaskOrchestrationExecutor } from "../taskorchestrationexecutor";
+import { TaskOrchestrationExecutor } from "../orchestrations/TaskOrchestrationExecutor";
 import { TaskID } from ".";
 import { DFTimerTask } from "./DFTimerTask";
